feat: ensure DeleQuestionUseCase delete a question from another user

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from '../repositories/question-respository';
 
 interface DeleteQuestionseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -9,11 +10,15 @@ interface DeleteQuestionseCaseResponse { }
 export class DeleteQuestionseCase {
   constructor(private questionrepository: QuestionsRepository) { }
 
-  async execute({ questionId }: DeleteQuestionseCaseRequest): Promise<DeleteQuestionseCaseResponse> {
+  async execute({ authorId, questionId }: DeleteQuestionseCaseRequest): Promise<DeleteQuestionseCaseResponse> {
     const question = await this.questionrepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found.')
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error('Not allowed.')
     }
 
     await this.questionrepository.delete(question)
